+++ src/test/java/com/stool/studentcooperationtools/websocket/controller/vote/VoteWebSocketControllerTest.java
@@ -1,71 +1,27 @@
 package com.stool.studentcooperationtools.websocket.controller.vote;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 import com.stool.studentcooperationtools.domain.vote.service.VoteService;
 import com.stool.studentcooperationtools.security.oauth2.dto.SessionMember;
 import com.stool.studentcooperationtools.websocket.CustomSessionHandlerAdapter;
+import com.stool.studentcooperationtools.websocket.WebsocketTestSupport;
 import com.stool.studentcooperationtools.websocket.controller.vote.request.VoteAddWebSocketRequest;
 import com.stool.studentcooperationtools.websocket.controller.vote.request.VoteDeleteSocketRequest;
 import com.stool.studentcooperationtools.websocket.controller.vote.response.VoteAddWebSocketResponse;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.messaging.converter.MappingJackson2MessageConverter;
-import org.springframework.messaging.simp.stomp.StompSession;
-import org.springframework.messaging.simp.stomp.StompSessionHandlerAdapter;
-import org.springframework.web.socket.WebSocketHttpHeaders;
-import org.springframework.web.socket.client.standard.StandardWebSocketClient;
-import org.springframework.web.socket.messaging.WebSocketStompClient;
-import org.springframework.web.socket.sockjs.client.SockJsClient;
-import org.springframework.web.socket.sockjs.client.WebSocketTransport;
 
-import java.util.List;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class VoteWebSocketControllerTest {
 
-    @LocalServerPort
-    private int port;
-    private String URL;
-    private StompSession stompSession;
-    private WebSocketStompClient stompClient;
+class VoteWebSocketControllerTest extends WebsocketTestSupport {
 
     @MockBean
     VoteService voteService;
-
-    @BeforeEach
-    void setUp() throws ExecutionException, InterruptedException, TimeoutException {
-        URL = "ws://localhost:%d/ws-stomp".formatted(port);
-        stompClient = new WebSocketStompClient(
-                new SockJsClient(List.of(new WebSocketTransport(new StandardWebSocketClient()))));
-        MappingJackson2MessageConverter messageConverter = new MappingJackson2MessageConverter();
-        ObjectMapper objectMapper = messageConverter.getObjectMapper();
-        objectMapper.registerModules(new JavaTimeModule(), new ParameterNamesModule());
-        stompClient.setMessageConverter(messageConverter);
-        stompSession = stompClient.connectAsync(URL,
-                new WebSocketHttpHeaders(),
-                new StompSessionHandlerAdapter() {
-                }).get(1, TimeUnit.SECONDS);
-    }
-
-    @AfterEach
-    void tearDown(){
-        if(stompSession.isConnected()){
-            stompSession.disconnect();
-        }
-    }
 
     @DisplayName("투표할 주제를 받아 투표를 추가한다.")
     @Test
